+++ R/CiteSource.R
@@ -51,4 +51,4 @@ utils::globalVariables(".")
 
 key_fields <- c("author", "title", "year", "journal", "abstract", "doi", "number", "pages",
                 "volume", "isbn", "record_id", "label", "source", "issue", "url",
-                "issn", "start_page", "ID")
+                "issn", "start_page", "ID", "type", "source_type")